Code Improvements to favorites viewpager
