+++ app/src/main/res/layout/fragment_dynamic_tab.xml
@@ -1,10 +1,13 @@
 <?xml version="1.0" encoding="utf-8"?>
+<layout
+    xmlns:app="http://schemas.android.com/apk/res-auto">
+    <data/>
 <androidx.constraintlayout.widget.ConstraintLayout
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".favorites.DynamicTabFragment">
+    tools:context=".favorites.screens.DynamicTabFragment">
 
     <TextView
         android:layout_width="match_parent"
@@ -12,3 +15,4 @@
         android:text="Blank Fragment" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
+</layout>
